Cleanup and added commentary
