+++ javaagents/src/main/java/g6Agent/perceptionAndMemory/Enties/StepAndMovement.java
@@ -1,0 +1,10 @@
+package g6Agent.perceptionAndMemory.Enties;
+
+/**
+ * @param step
+ * @param movement
+ *
+ * @author Kai Müller
+ */
+public record StepAndMovement(int step, Movement movement) {
+    }
