+++ javaagents/src/main/java/g6Agent/perceptionAndMemory/Interfaces/AgentVisionReporter.java
@@ -2,24 +2,44 @@ package g6Agent.perceptionAndMemory.Interfaces;
 
 import g6Agent.perceptionAndMemory.Enties.AgentNameAndPosition;
 import g6Agent.perceptionAndMemory.Enties.Block;
+import g6Agent.perceptionAndMemory.Enties.Vision;
 import g6Agent.services.Point;
 
 import java.util.List;
 
 /**
- * NOT IMPLEMENTED!
- * <p>
  * Interface to syncronize the vision of this agent with the vision of other Agents
+ *
+ * @author Kai Müller
  */
 public interface AgentVisionReporter {
 
-    void reportMyVision(List<Block> dispensers, List<Block> blocks, List<Point> roleZones, List<Point> goalZones, List<Point> obstacles);
+    /**
+     * report what the vision of this Agent to other Agents
+     * @param vision the Vision
+     */
+    void reportMyVision(Vision vision);
 
+    /**
+     * updates the Vision
+     */
     void updateMyVisionWithSightingsOfOtherAgents();
 
+    /**
+     * use once per step, does its routine for this step
+     */
     void handleStep();
 
+    /**
+     *
+     * @return the Positions and Names of the Known Agents
+     */
     List<AgentNameAndPosition> getKnownAgentPositions();
 
+    /**
+     * Returns the Position of an Agent with a given Name, null if is unknown
+     * @param agentname the Name of the Agent
+     * @return the position, null if is unknown
+     */
     Point getPositionOfAgent(String agentname);
 }
