+++ javaagents/src/main/java/g6Agent/perceptionAndMemory/SwarmSightController.java
@@ -2,40 +2,41 @@ package g6Agent.perceptionAndMemory;
 
 import eis.iilang.*;
 import g6Agent.MailService;
-import g6Agent.perceptionAndMemory.Enties.AgentNameAndPosition;
-import g6Agent.perceptionAndMemory.Enties.Block;
-import g6Agent.perceptionAndMemory.Enties.LastActionMemory;
-import g6Agent.perceptionAndMemory.Enties.Movement;
+import g6Agent.perceptionAndMemory.Enties.*;
+import g6Agent.perceptionAndMemory.messages.IntroductionRequest;
+import g6Agent.perceptionAndMemory.messages.KnownAgentsNotificationMessage;
+import g6Agent.perceptionAndMemory.messages.MovementNotificationMessage;
+import g6Agent.perceptionAndMemory.messages.MyVisionMessage;
 import g6Agent.perceptionAndMemory.Interfaces.*;
 import g6Agent.services.Direction;
 import g6Agent.services.Point;
+import g6Agent.services.SpeedCalculator;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.*;
 
 /**
  * Controller for determining the relative Positions of other Agents and Communicating their Vision
+ *
+ * @author Kai Müller
  */
 class SwarmSightController implements LastActionListener, CommunicationModuleSwarmSightControllerInterface, AgentVisionReporter {
 
     private final MailService mailservice;
     private final PerceptionAndMemory perceptionAndMemory;
-
-    private final PerceptionAndMemoryInput perceptionAndMemoryInput;
     private final HashMap<String, StepAndMovement> attemptedMovements;
-
     private final SwarmSightModel swarmSightModel;
     private final String agentname;
-
-    int messageCounter;
+    private final SwarmSightDataConverter dataConverter;
     private List<IntroductionRequest> requestsToAnswer;
     private List<IntroductionRequest> acceptMessagesThisStep;
-    private final HashMap<String, Vison> visions;
+    private final HashMap<String, Vision> visions;
+    int messageCounter;
 
     SwarmSightController(MailService mailservice, PerceptionAndMemory perceptionAndMemory, PerceptionAndMemoryInput perceptionAndMemoryInput, String agentname) {
         this.mailservice = mailservice;
         this.perceptionAndMemory = perceptionAndMemory;
-        this.perceptionAndMemoryInput = perceptionAndMemoryInput;
+        this.dataConverter = new SwarmSightDataConverter(perceptionAndMemoryInput, perceptionAndMemory);
         this.swarmSightModel = new SwarmSightModel();
         this.agentname = agentname;
         this.attemptedMovements = new HashMap<>();
@@ -45,23 +46,22 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
         this.visions = new HashMap<>();
     }
 
-
     @Override
     public void processMovementNotification(Percept message, String sender) {
         if (!sender.equals(agentname)) {
-            int step = ((Numeral) message.getParameters().get(1)).getValue().intValue();
+            MovementNotificationMessage notificationMessage = MovementNotificationMessage.fromMail(message, sender);
+            if(notificationMessage == null) return;
             attemptedMovements.remove(sender);
             Movement movement = new Movement(
-                    parameterlistToListOfDirections((ParameterList) message.getParameters().get(2)), //directions
-                    ((Numeral) message.getParameters().get(3)).getValue().intValue()                //speed
+                    parameterlistToListOfDirections((ParameterList) notificationMessage.movementParameter()), //directions
+                    notificationMessage.speed()        //speed
             );
             swarmSightModel.notifiedOfMovement(sender, movement);
         }
-
     }
 
     @Override
-    public void broadcastActionAttempt(Action action) {
+    public void broadcastActionAttempt(@NotNull Action action) {
         if (action.getName().equals("move")) {
             ParameterList parameters = new ParameterList(action.getParameters());
 
@@ -69,13 +69,13 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
                     new Numeral(messageCounter),                        //Clock
                     new Numeral(perceptionAndMemory.getCurrentStep()), // Step
                     parameters,                                        // Directions
-                    new Numeral(determineyourOwnSpeed())                // Speed
+                    new Numeral(SpeedCalculator.calculateSpeed(perceptionAndMemory)) // Speed
             ), agentname);
         }
     }
 
     @Override
-    public void processMovementAttempt(Percept message, String sender) {
+    public void processMovementAttempt(@NotNull Percept message, String sender) {
         if (message.getName().equals("MOVEMENT_ATTEMPT")) {
             if (!sender.equals(agentname)) {
                 attemptedMovements.put(sender, new StepAndMovement(
@@ -87,7 +87,7 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
     }
 
     @NotNull
-    private List<Direction> parameterlistToListOfDirections(ParameterList directionIdentifiers) {
+    private List<Direction> parameterlistToListOfDirections(@NotNull ParameterList directionIdentifiers) {
         List<Direction> directions = new ArrayList<>();
         for (Parameter p : directionIdentifiers) {
             directions.add(Direction.fromIdentifier((Identifier) p));
@@ -96,11 +96,11 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
     }
 
     @Override
-    public void processIntroductionRequest(Percept message, String sender) {
+    public void processIntroductionRequest(@NotNull Percept message, String sender) {
         if (message.getName().equals("INTRODUCTION_REQUEST")) {
             if (!sender.equals(agentname)) {
                 IntroductionRequest request = IntroductionRequest.fromMail(message, sender);
-                if (request.step == perceptionAndMemory.getCurrentStep()) {
+                if (request.step() == perceptionAndMemory.getCurrentStep()) {
                     answerIntroductionRequest(request);
                 } else {
                     this.requestsToAnswer.add(request);
@@ -111,309 +111,70 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
 
     private void answerIntroductionRequest(IntroductionRequest request) {
         for (Point agentposition : perceptionAndMemory.getFriendlyAgents()) {
-            if (agentposition.equals(request.position)) {
+            if (agentposition.equals(request.position())) {
                 request.sendAccept(mailservice, this.agentname);
             }
         }
     }
 
     @Override
-    public void processIntroductionAccept(Percept message, String sender) {
+    public void processIntroductionAccept(@NotNull Percept message, String sender) {
         if (message.getName().equals("INTRODUCTION_ACCEPT")) {
             IntroductionRequest acceptMessage = IntroductionRequest.fromMail(message, sender);
             this.acceptMessagesThisStep.add(acceptMessage);
         }
     }
 
-
     @Override
-    public void reportLastAction(LastActionMemory lastAction) {
+    public void reportLastAction(@NotNull LastActionMemory lastAction) {
         //If moved at all
         if (lastAction.getName().equals("move")) {
-            int speed = determineSpeedOfLastAction(lastAction);
+            int speed = SpeedCalculator.determineSpeedOfLastAction(lastAction, perceptionAndMemory);
             if (speed > 0) {
-
+                MovementNotificationMessage movementNotificationMessage
+                        = new MovementNotificationMessage(
+                                messageCounter,
+                                perceptionAndMemory.getCurrentStep(),
+                                lastAction.getParameters().get(0),
+                                speed,
+                                agentname);
                 List<Direction> directions = parameterlistToListOfDirections((ParameterList) lastAction.getParameters().get(0));
                 Movement movement = new Movement(directions, speed);
                 swarmSightModel.movedMyself(movement);
-                mailservice.broadcast(new Percept("MOVEMENT_NOTIFICATION",
-                        new Numeral(messageCounter),
-                        new Numeral(perceptionAndMemory.getCurrentStep()),
-                        lastAction.getParameters().get(0),
-                        new Numeral(speed)
-                ), agentname);
+                movementNotificationMessage.broadcast(mailservice);
             }
         }
     }
 
-    private int determineSpeedOfLastAction(LastActionMemory lastAction) {
-        if (!(lastAction.getSuccessMessage().equals("success")
-                || lastAction.getSuccessMessage().equals("partial_success"))) {
-            return 0;
-        }
-
-        if (lastAction.getSuccessMessage().equals("partial_success")) {
-            return 1;
-            //TODO Explorer for whom it could be 1 or 2 is unhandeld
-        } else {
-            return determineyourOwnSpeed();
-        }
-    }
-
-    private int determineyourOwnSpeed() {
-        if (perceptionAndMemory.getCurrentRole() == null) return 1;
-        int speed;
-        if (perceptionAndMemory.getCurrentRole().getMovementSpeed().size() <= perceptionAndMemory.getDirectlyAttachedBlocks().size()) {  //logic to get array entry of speed
-            speed = 0;
-        } else {
-            speed = perceptionAndMemory.getCurrentRole().getMovementSpeed().get(perceptionAndMemory.getDirectlyAttachedBlocks().size());
-        }
-        return speed;
-    }
-
     @Override
-    public void processVisionNotification(Percept message, String sender) {
+    public void processVisionNotification(@NotNull Percept message, String sender) {
         if (message.getName().equals("MY_VISION")) {
             if (!sender.equals(agentname)) {
                 if (swarmSightModel.isKnown(sender)) {
-                    List<Block> dispensers = new ArrayList<>();
-                    List<Block> blocks = new ArrayList<>();
-                    List<Point> roleZones = new ArrayList<>();
-                    List<Point> goalZones = new ArrayList<>();
-                    List<Point> obstacles = new ArrayList<>();
-                    //parse Percept and write to lists, save as vision
-                    for (Parameter parameter : (ParameterList) message.getParameters().get(0)) {
-                        if (parameter instanceof Function function) {
-                            switch (function.getName()) {
-                                case "dispenser" -> dispensers.add(new Block(
-                                        new Point(
-                                                ((Numeral) function.getParameters().get(1)).getValue().intValue(),
-                                                ((Numeral) function.getParameters().get(2)).getValue().intValue()
-                                        ),
-                                        ((Identifier) function.getParameters().get(0)).toProlog()
-                                ));
-                                case "block" -> blocks.add(new Block(
-                                        new Point(
-                                                ((Numeral) function.getParameters().get(1)).getValue().intValue(),
-                                                ((Numeral) function.getParameters().get(2)).getValue().intValue()
-                                        ),
-                                        ((Identifier) function.getParameters().get(0)).toProlog()
-                                ));
-                                case "rolezone" -> roleZones.add(new Point(
-                                        ((Numeral) function.getParameters().get(0)).getValue().intValue(),
-                                        ((Numeral) function.getParameters().get(1)).getValue().intValue()
-                                ));
-                                case "goalzone" -> goalZones.add(new Point(
-                                        ((Numeral) function.getParameters().get(0)).getValue().intValue(),
-                                        ((Numeral) function.getParameters().get(1)).getValue().intValue()
-                                ));
-                                case "obstacle" -> obstacles.add(new Point(
-                                        ((Numeral) function.getParameters().get(0)).getValue().intValue(),
-                                        ((Numeral) function.getParameters().get(1)).getValue().intValue()
-                                ));
-                            }
-                        }
+                    MyVisionMessage receivedVisionMessage = MyVisionMessage.fromMail(message, sender);
+                    if (receivedVisionMessage != null) {
+                        this.visions.put(sender, receivedVisionMessage.vision());
                     }
-                    this.visions.put(sender, new Vison(dispensers, blocks, roleZones, goalZones, obstacles));
                 }
             }
         }
     }
 
-
-    record Vison(List<Block> dispensers, List<Block> blocks, List<Point> roleZones, List<Point> goalZones,
-                 List<Point> obstacles) {
-    }
-
     @Override
-    public void reportMyVision(List<Block> dispensers, List<Block> blocks, List<Point> roleZones, List<Point> goalZones, List<Point> obstacles) {
-        ParameterList functions = new ParameterList();
-        for (Block dispenser : dispensers) {
-            functions.add(new Function("dispenser", new Identifier(dispenser.getBlocktype()), new Numeral(dispenser.getCoordinates().x), new Numeral(dispenser.getCoordinates().y)));
-        }
-
-        for (Block dispenser : dispensers) {
-            functions.add(new Function("block", new Identifier(dispenser.getBlocktype()), new Numeral(dispenser.getCoordinates().x), new Numeral(dispenser.getCoordinates().y)));
-        }
-
-        for (Point roleZone : roleZones) {
-            functions.add(new Function("rolezone", new Numeral(roleZone.getX()), new Numeral(roleZone.getY())));
-        }
-        for (Point goalZone : goalZones) {
-            functions.add(new Function("goalzone", new Numeral(goalZone.x), new Numeral(goalZone.y)));
-        }
-        for (Point obstacle : obstacles) {
-            functions.add(new Function("obstacle", new Numeral(obstacle.x), new Numeral(obstacle.y)));
-        }
-        mailservice.broadcast(new Percept("MY_VISION", functions), agentname);
+    public void reportMyVision(Vision vision) {
+        MyVisionMessage myVisionMessage = new MyVisionMessage(vision, agentname);
+        myVisionMessage.broadcast(mailservice);
     }
 
     @Override
     public void updateMyVisionWithSightingsOfOtherAgents() {
-        HashSet<Block> uniqueDispensers = new HashSet<>();
-        HashSet<Block> uniqueBlocks = new HashSet<>();
-        HashSet<Point> uniqueGoalZones = new HashSet<>();
-        HashSet<Point> uniqueRoleZones = new HashSet<>();
-        HashSet<Point> uniqueObstacles = new HashSet<>();
-        for (AgentNameAndPosition agent : swarmSightModel.knownAgents()) {
-            if (agent.position() != null) {
-                Vison vison = visions.get(agent.name());
-                if (vison != null) {
-                    addVisionToHashSets(uniqueDispensers, uniqueBlocks, uniqueGoalZones, uniqueRoleZones, uniqueObstacles, agent, vison);
-                }
-            }
-        }
-        flushHashSetsToPerceptionAndMemory(uniqueDispensers, uniqueBlocks, uniqueGoalZones, uniqueRoleZones, uniqueObstacles);
-    }
-
-    private void flushHashSetsToPerceptionAndMemory(HashSet<Block> uniqueDispensers, HashSet<Block> uniqueBlocks, HashSet<Point> uniqueGoalZones, HashSet<Point> uniqueRoleZones, HashSet<Point> uniqueObstacles) {
-        for (Block dispenser : uniqueDispensers) {
-            try {
-                addDispenserToPerception(dispenser);
-            } catch (Exception e) {
-                System.out.println(" IN UPDATE VISION : Problem with adding dispenser");
-                e.printStackTrace();
-            }
-        }
-        for (Block block : uniqueBlocks) {
-            try {
-                addBlockToPerception(block);
-            } catch (Exception e) {
-                System.out.println(" IN UPDATE VISION : Problem with adding block");
-                e.printStackTrace();
-            }
-        }
-        for (Point obstacle : uniqueObstacles) {
-            try {
-                addObstacleToPerception(obstacle);
-            } catch (Exception e) {
-                System.out.println(" IN UPDATE VISION : Problem with adding obstacle");
-                e.printStackTrace();
-            }
-        }
-        for (Point goalzone : uniqueGoalZones) {
-            try {
-                addGoalZoneToPerception(goalzone);
-            } catch (Exception e) {
-                System.out.println(" IN UPDATE VISION : Problem with adding goalZone");
-                e.printStackTrace();
-            }
-        }
-        for (Point roleZone : uniqueRoleZones) {
-            try {
-                addRoleZoneToPerception(roleZone);
-            } catch (Exception e) {
-                System.out.println(" IN UPDATE VISION : Problem with adding roleZone");
-                e.printStackTrace();
-            }
-        }
-    }
-
-    private void addVisionToHashSets(HashSet<Block> uniqueDispensers, HashSet<Block> uniqueBlocks, HashSet<Point> uniqueGoalZones, HashSet<Point> uniqueRoleZones, HashSet<Point> uniqueObstacles, AgentNameAndPosition agent, Vison vison) {
-        //add dispensers to uniqueDispensers
-        for (Block dispenser : vison.dispensers()) {
-            Block dispenserWithNewPosition =
-                    new Block(dispenser.getCoordinates().add(agent.position()), dispenser.getBlocktype());
-            if (isOutOfSight(dispenserWithNewPosition)) {
-                uniqueDispensers.add(dispenserWithNewPosition);
-            }
-        }
-        //add blocks to uniqueBlocks
-        for (Block block : vison.blocks()) {
-            Block blockWithNewPosition =
-                    new Block(block.getCoordinates().add(agent.position()), block.getBlocktype());
-            if (isOutOfSight(blockWithNewPosition)) {
-                uniqueBlocks.add(blockWithNewPosition);
-            }
-        }
-        //add goalzones to uniqueGoalzones
-        for (Point goalZone : vison.goalZones()) {
-            Point goalZoneWithNewPosition = goalZone.add(agent.position());
-            if (isOutOfSight(goalZoneWithNewPosition)) {
-                uniqueGoalZones.add(goalZoneWithNewPosition);
-            }
-        }
-        //add rolezones to uniqueRoleZones
-        for (Point roleZone : vison.roleZones()) {
-            Point roleZoneWithNewPosition = roleZone.add(agent.position());
-            if (isOutOfSight(roleZoneWithNewPosition)) {
-                uniqueRoleZones.add(roleZoneWithNewPosition);
-            }
-        }
-        //add obstacles to uniqueObstacles
-        for (Point obstacle : vison.obstacles()) {
-            Point obstacleWithNewPosition = obstacle.add(agent.position());
-            if (isOutOfSight(obstacleWithNewPosition)) {
-                uniqueObstacles.add(obstacleWithNewPosition);
-            }
-        }
-    }
-
-    private void addRoleZoneToPerception(Point roleZone) throws Exception {
-        perceptionAndMemoryInput.handleRoleZone(new Percept(
-                "roleZone",
-                new Numeral(roleZone.x),
-                new Numeral(roleZone.y)
-        ));
-    }
-
-    private void addGoalZoneToPerception(Point goalzone) throws Exception {
-        perceptionAndMemoryInput.handleGoalZone(new Percept(
-                "goalZone",
-                new Numeral(goalzone.x),
-                new Numeral(goalzone.y)
-        ));
-    }
-
-    private void addObstacleToPerception(Point obstacle) throws Exception {
-        perceptionAndMemoryInput.handleThingPercept(
-                new Percept("thing",
-                        new Numeral(obstacle.x),
-                        new Numeral(obstacle.y),
-                        new Identifier("obstacle"),
-                        new Identifier("")
-                )
-        );
-    }
-
-    private void addBlockToPerception(Block block) throws Exception {
-        perceptionAndMemoryInput.handleThingPercept(
-                new Percept("thing",
-                        new Numeral(block.getCoordinates().x),
-                        new Numeral(block.getCoordinates().y),
-                        new Identifier("block"),
-                        new Identifier(block.getBlocktype())
-                )
-        );
-    }
-
-
-    private void addDispenserToPerception(Block dispenser) throws Exception {
-        perceptionAndMemoryInput.handleThingPercept(
-                new Percept("thing",
-                        new Numeral(dispenser.getCoordinates().x),
-                        new Numeral(dispenser.getCoordinates().y),
-                        new Identifier("dispenser"),
-                        new Identifier(dispenser.getBlocktype())
-                )
-        );
-    }
-
-    private boolean isOutOfSight(Point point) {
-        if (perceptionAndMemory.getCurrentRole() == null) return false; // if no Role ignore
-        return point.manhattanDistanceTo(new Point(0, 0)) > perceptionAndMemory.getCurrentRole().getVisionRange();
-    }
-
-    private boolean isOutOfSight(Block block) {
-        if (perceptionAndMemory.getCurrentRole() == null) return false; // if no Role ignore
-        return block.getCoordinates().manhattanDistanceTo(new Point(0, 0)) > perceptionAndMemory.getCurrentRole().getVisionRange();
+        dataConverter.updateMyVisionWithSightingsOfOtherAgents(swarmSightModel, visions);
     }
 
     @Override
     public void handleStep() {
-        handleUnanseweredRequests();
+        handleUnansweredRequests();
         this.requestsToAnswer = new ArrayList<>();
-
         //check accepts, each accept, which is a one of -> add sighting
         compareRequestsSendWithAccepts();
         this.acceptMessagesThisStep = new ArrayList<>();
@@ -425,30 +186,20 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
     private void broadcastKnownAgents() {
         List<AgentNameAndPosition> knownAgents = swarmSightModel.knownAgents();
         if (!knownAgents.isEmpty()) {
-            ParameterList listOfKnownAgentPercepts = new ParameterList();
-            for (AgentNameAndPosition agent : knownAgents) {
-                if (!agent.name().equals(agentname)) {
-                    listOfKnownAgentPercepts.add(new Function(agent.name(), new Numeral(agent.position().y), new Numeral(agent.position().y)));
-                }
-            }
-            mailservice.broadcast(new Percept("KNOWN_AGENTS", listOfKnownAgentPercepts), agentname);
+            KnownAgentsNotificationMessage knownAgentsMessage = new KnownAgentsNotificationMessage(knownAgents, this.agentname);
+            knownAgentsMessage.brodacast(mailservice);
         }
     }
 
     @Override
-    public void processKnownAgentsNotification(Percept message, String sender) {
-        if (message.getName().equals("KNOWN_AGENTS") && swarmSightModel.isKnown(sender)) {
-            for (Parameter parameter : (ParameterList) message.getParameters().get(0)) {
-                if (parameter instanceof Function function) {
-                    String name = function.getName();
-                    //if agent is not known add to known Agents with relative Position to sender
-                    if (!swarmSightModel.isKnown(name) && !name.equals(agentname)) {
-                        Point position = new Point(
-                                ((Numeral) function.getParameters().get(0)).getValue().intValue(),
-                                ((Numeral) function.getParameters().get(1)).getValue().intValue()
-                        );
-                        swarmSightModel.heardOfAgentPosition(name, position, sender);
-                    }
+    public void processKnownAgentsNotification(@NotNull Percept message, String sender) {
+        if (message.getName().equals(KnownAgentsNotificationMessage.identifier()) && swarmSightModel.isKnown(sender)) {
+            KnownAgentsNotificationMessage knownAgentsMessage = KnownAgentsNotificationMessage.fromMail(message, sender);
+            if (knownAgentsMessage == null) return;
+            for (AgentNameAndPosition agentInfo : knownAgentsMessage.knownAgents()) {
+                //if agent is not known add to known Agents with relative Position to sender
+                if (!swarmSightModel.isKnown(agentInfo.name()) && !agentInfo.name().equals(agentname)) {
+                    swarmSightModel.heardOfAgentPosition(agentInfo.name(), agentInfo.position(), sender);
                 }
             }
         }
@@ -464,9 +215,9 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
         return swarmSightModel.getAgentPosition(agentname);
     }
 
-    private void handleUnanseweredRequests() {
+    private void handleUnansweredRequests() {
         for (IntroductionRequest request : requestsToAnswer) {
-            if (request.step == perceptionAndMemory.getCurrentStep()) {
+            if (request.step() == perceptionAndMemory.getCurrentStep()) {
                 answerIntroductionRequest(request);
             }
         }
@@ -507,7 +258,6 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
         }
     }
 
-
     private void checkForOtherAgents() {
         List<Point> unknownAgents = new ArrayList<>();
         //determine known Agents in sight
@@ -526,7 +276,6 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
         }
     }
 
-
     private boolean comparePositionsInInternalMap(Point agentPosition) {
         for (AgentNameAndPosition agentInMemory : swarmSightModel.knownAgents()) {
             //if movement is known
@@ -539,12 +288,12 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
         return false;
     }
 
-    private boolean comparePositionWithMovementVectors(Point agentPosition, AgentNameAndPosition agentInMemory) {
+    private boolean comparePositionWithMovementVectors(Point agentPosition, @NotNull AgentNameAndPosition agentInMemory) {
         boolean isIdentified = false;
         StepAndMovement attemptedMove = attemptedMovements.get(agentInMemory.name());
         if (attemptedMove != null) {
             Point positionBefore = agentInMemory.position();
-            Point positionAfter = agentInMemory.position().add(attemptedMove.movement.asVector());
+            Point positionAfter = agentInMemory.position().add(attemptedMove.movement().asVector());
             if (positionBefore.x == positionAfter.x) {
                 int smallerY = (Math.min(positionBefore.y, positionAfter.y));
                 int biggerY = (Math.max(positionBefore.y, positionAfter.y));
@@ -564,50 +313,5 @@ class SwarmSightController implements LastActionListener, CommunicationModuleSwa
             }
         }
         return isIdentified;
-    }
-
-    private record StepAndMovement(int step, Movement movement) {
-    }
-
-    /**
-     * An request for if a agent is seen.
-     *
-     * @param clock    the clock
-     * @param step     the step
-     * @param position the inverted position from the sender. is the position the reciever should see the sender at.
-     * @param sender   name of the sender
-     */
-    private record IntroductionRequest(int clock, int step, Point position, String sender) {
-
-
-        void broadcast(MailService mailservice) {
-            mailservice.broadcast(new Percept("INTRODUCTION_REQUEST",
-                    new Numeral(this.clock), new Numeral(this.step), new Numeral(this.position.x), new Numeral(this.position.y)
-            ), this.sender);
-        }
-
-        static IntroductionRequest fromMail(Percept percept, String sender) {
-            return new IntroductionRequest(
-                    ((Numeral) percept.getParameters().get(0)).getValue().intValue(), //clock
-                    ((Numeral) percept.getParameters().get(1)).getValue().intValue(), //step
-                    new Point(
-                            ((Numeral) percept.getParameters().get(2)).getValue().intValue(),
-                            ((Numeral) percept.getParameters().get(3)).getValue().intValue()
-                    ),      //position
-                    sender //sender
-            );
-        }
-
-
-        void sendAccept(MailService mailService, String sender) {
-            mailService.sendMessage(
-                    new Percept(
-                            "INTRODUCTION_ACCEPT",
-                            new Numeral(this.clock), new Numeral(this.step), new Numeral(this.position.x), new Numeral(this.position.y)
-                    ),
-                    this.sender,
-                    sender
-            );
-        }
     }
 }
