+++ javaagents/src/main/java/g6Agent/services/Direction.java
@@ -1,7 +1,6 @@
 package g6Agent.services;
 
 import eis.iilang.Identifier;
-import lombok.EqualsAndHashCode;
 
 import java.util.LinkedList;
 import java.util.List;
